+++ Changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- Unit tests for data_file
 - Linting and Formatting to GitHub Actions
 
 ## Fixed